+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* Fixed File Content type
+
 ## v0.3.0 (2024-05-21)
 * Added Message model structures
 * Added Suggestions to Repository