+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## v0.3.2 (2024-05-21)
+* Fixed Message jsonSerialize
+* Fixed composer dev version
+
 ## v0.3.1 (2024-05-21)
 * Fixed File Content type